backend/requirements.txt
django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
django-cors-headers>=4.3
django-redis>=5.4
psycopg2-binary>=2.9
cryptography>=42.0
python-decouple>=3.8
dj-database-url>=2.2
gunicorn>=23.0
whitenoise>=6.8
